+++ packages/connect-multichain/src/domain/multichain/index.ts
@@ -12,7 +12,7 @@ import type { StoreClient } from '../store/client';
 import type { InvokeMethodOptions, RPCAPI, Scope } from './api/types';
 import type { MultichainOptions, ExtendedTransport } from './types';
 
-export type SDKState =
+export type ConnectionStatus =
   | 'pending'
   | 'loaded'
   | 'disconnected'
@@ -34,7 +34,7 @@ export enum TransportType {
 export abstract class MultichainCore extends EventEmitter<SDKEvents> {
   abstract storage: StoreClient;
 
-  abstract state: SDKState;
+  abstract status: ConnectionStatus;
 
   abstract provider: MultichainApiClient<RPCAPI>;